server/requirements.txt
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.3.0
langchain-ollama>=0.3.0
langchain-community>=0.3.0
langgraph>=0.4.0
fastapi>=0.115.0
uvicorn>=0.34.0
pydantic>=2.10.0
python-dotenv>=1.1.0
httpx>=0.28.0
psycopg[binary]>=3.2.0
